+++ .circleci/config.yml
@@ -151,6 +151,7 @@ jobs:
             export DRIVE_FOLDER_NAME=gke-stockprice-dump-test
             export DRIVE_PERMISSION_GMAIL=${GMAIL}
             go test -v ./... -tags integration -timeout 30m
+          # TODO： integration_test後にclusterが無くなっているか確認するステップを入れる
 
   create_gke_cluster:
     working_directory: ~/go/src/github.com/ludwig125 # この設定によって、stepsのcheckout時にworking_directory以下にgke-stockpriceリポジトリがgit cloneされる
