+++ cloudsql/cloudsql_test.go
@@ -1,6 +1,6 @@
 // +build integration
 
-package gcloud
+package cloudsql
 
 import (
 	"testing"
